[BOLT] Add support for safe-icf
Summary:
Identical Code Folding (ICF) folds functions that are identical into one function, and updates symbol addresses to the new address. This reduces the size of a binary, but can lead to problems. For example when function pointers are compared. This can be done either explicitly in the code or generated IR by optimization passes like Indirect Call Promotion (ICP). After ICF what used to be two different addresses become the same address. This can lead to a different code path being taken.

This is where safe ICF comes in. Linker (LLD) does it using address significant section generated by clang. If symbol is in it, or an object doesn't have this section symbols are not folded.

BOLT does not have the information regarding which objects do not have this section, so can't re-use this mechanism.

This implementation scans code section and conservatively marks functions symbols as unsafe. It treats symbols as unsafe if they are used in non-control flow instruction. It also scans through the data relocation sections and does the same for relocations that reference a function symbol. The latter handles the case when function pointer is stored in a local or global variable, etc. If a relocation address points within a vtable these symbols are skipped.

Test Plan: ninja check-all

Reviewers: aaupov, #llvm-bolt

Reviewed By: aaupov

Differential Revision: https://phabricator.intern.facebook.com/D65799965

Tasks: T180058628

Tags: accept2ship