[BOLT] Flow conservation metrics
Summary:
Add two additional profile quality stats for CG (call graph) and CFG (control flow graph) flow conservations besides the CFG discontinuity stats. The two new stats quantify how different "in-flow" is from "out-flow" in the following cases where they should be equal. The smaller the reported stats, the better the flow conservations are.

CG flow conservation: for each function that is not a program entry, the number of times the function is called according to CG ("in-flow") should be equal to the number of times the transition from an entry basic block of the function to another basic block within the function is recorded ("out-flow").

CFG flow conservation: for each basic block that is not a function entry or exit, the number of times the transition into this basic block from another basic block within the function is recorded ("in-flow") should be equal to the number of times the transition from this basic block to another basic block within the function is recorded ("out-flow").

Use -v=1 for more detailed bucketed stats, and use -v=2 to dump functions / basic blocks with bad flow conservations.

Test Plan:
`bolt/test/X86/profile-quality-reporting.test`
`wrapper-script.test` failure is handled in D70419608

Reviewers: aaupov, xuanzh

Reviewed By: aaupov

Differential Revision: https://phabricator.intern.facebook.com/D68176994