[server-llvm-19][C++20][Modules] Fix crash/compiler error due broken AST links (#123648)
Summary:
Backport of: https://github.com/llvm/llvm-project/pull/123648

This PR fixes bugreport
https://github.com/llvm/llvm-project/issues/122493 The root problem is
the same as before lambda function and DeclRefExpr references a variable
that does not belong to the same module as the enclosing function body.
Therefore iteration over the function body doesn't visit the VarDecl.
Before this change RelatedDeclsMap was created only for canonical decl
but in reality it has to be done for the definition of the function that
does not always match the canonical decl.

Test Plan: check-clang

Reviewers: wenlei, matthiasb, #compiler_toolchain

Reviewed By: wenlei

Differential Revision: https://phabricator.intern.facebook.com/D68623387