[server-llvm-19][clang][modules] Fix lambda and its enclosing function are not loaded from same module. (#142467)
Summary:
Backport of: https://github.com/llvm/llvm-project/pull/142467

This is a follow-up fix to
https://github.com/llvm/llvm-project/pull/109167.

Previously, we stored a mapping between the enclosing function and the
lambda class declaration. When loading the enclosing function, we also
loaded the corresponding lambda class declaration. However, loading the
lambda class declaration does not guarantee that its call operator (a
`CXXMethodDecl`) is loaded as well.
As a result, if the lambda call operator is later loaded from a
different module, we can end up with a `DeclRefExpr` that refers to a
`VarDecl` from a different module — leading to inconsistencies.

To fix this, we should ensure the lambda call operator itself is loaded.

Fixes #141582

Test Plan: check-clang

Rollback Plan:

Reviewers: ych, #compiler_toolchain, nml, wenlei

Reviewed By: ych

Differential Revision: https://phabricator.intern.facebook.com/D76587997

Tags: accept2ship