[BOLT][AArch64] Always cover veneers in lite mode
Summary: If a veneer is not disassembled in lite mode, the veneer elimination pass will not recognize it as such and the call will remain unchanged. Later, we may need to insert a new veneer for such code ending up with a veneer to veneer. To avoid such suboptimal code generation, always disassemble veneers and guarantee that they are converted to direct calls in BOLT.

Test Plan: SC

Rollback Plan:

Reviewers: aaupov, #llvm-bolt, shatian

Reviewed By: aaupov

Differential Revision: https://phabricator.intern.facebook.com/D76486777

Tags: aarch64