[BOLT] Fix handling of ambiguous jump table entries
Summary:
Jump tables may contain entries that point immediately past the end of
their parent function. Normally, such entries are generated by the
compiler as a result of builtin_unreachable() case. We used to replace
those entries with a label belonging to their parent function assuming
the destination doesn't matter if it's an undefined behavior.

However, if such entry is at the end of the jump table, it could be a
real function pointer, not a jump table entry. We rely on heuristics to
detect such cases and can drop the trailing function pointer entries
from the table.

The problem presents when the "unreachable" ambiguous entry is followed
by another ambiguous entry corresponding to the start of the parent
function. In this case we accept pointers as entries and may incorrectly
update the function pointer.

The solution is to keep ambiguous "unreachable" jump table entries
identical to the original input, i.e. point to the same function.
This change does not affect CFG, but results in the entries being
updated with the new function address if it gets relocated.

Test Plan: test case added

Rollback Plan:

Reviewers: rafaelauler, #llvm-bolt, ottoni

Reviewed By: rafaelauler

Differential Revision: https://phabricator.intern.facebook.com/D80288960

Tasks: T231637331