[HOTFIX] Revert "[BOLT][AArch64] Fix another cause of extra entry point misidentification (#155055)"
This reverts commit a9c1ae8672613ca479724dc37c8732aeca7d9bf8.

Maksim:
According to this patch https://reviews.llvm.org/D127111, we do allow
data at the start of a function on ARM. That contradicts the assembly
generated by lua jit, and a common sense to some extend (after all, you
are expect to be able to call a function). The fix I had in mind for
T238531789 breaks the test from this patch. For now, let's revert the
change with the assertion in "Rouen" release.