Rebase: [BOLT][AArch64] Add experimental call relaxation pass
Summary:
Add a function call relaxation pass that groups functions into clusters,
each with a maximum size of 128MB. This pass is enabled using the
--relax-exp` option.

* Within each cluster: Function calls do not require relaxation.

* Between clusters: thunks are created at the cluster boundaries to
handle inter-cluster calls. These thunks can be either short or long,
depending on the proximity of the thunk destination. When possible,
a long thunk may be shared by both adjacent clusters.

* Hot functions: If all hot functions (after reordering) fit within a
single cluster (under 128MB), no thunks are needed on the hot path, as
determined by the profile.

* PLT behavior: For the Procedure Linkage Table (PLT), it is currently
assumed that the hottest cluster will be placed adjacent to the PLT, so
thunks are not required for that cluster.

* Pass behavior: When enabled, this pass replaces the LongJmp pass. It
typically runs much faster because it does not attempt to predict the
exact code layout.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D89636184

Test Plan: SC

Rollback Plan: 

Reviewers: #llvm-bolt

Differential Revision: https://phabricator.intern.facebook.com/D90035566